Rename write_file_handle.
