+++ rust/gitxetcore/src/xetblob/file_operations/batched_wfile_handle.rs
