+++ rust/gitxetcore/src/xetblob/file_operations/mod.rs
@@ -1,9 +1,9 @@
 mod batch_operations;
+mod batched_wfile_handle;
 mod cp_operations;
 mod fs_interface;
-mod write_file_handle;
 mod write_transaction_wrapper;
 
 pub use batch_operations::XetRepoOperationBatch;
+pub use batched_wfile_handle::BatchedWriteFileHandle;
 pub use cp_operations::perform_copy;
-pub use write_file_handle::BatchedWriteFileHandle;
